refactor: 로또 생성 방법 전략패턴으로 변경
