+++ src/main/java/lotto/domain/GenerateLottoNumberStrategy.java
@@ -1,0 +1,8 @@
+package lotto.domain;
+
+import java.util.List;
+
+@FunctionalInterface
+public interface GenerateLottoNumberStrategy {
+    List<Integer> generate();
+}
