+++ src/main/java/lotto/domain/strategy/ManualLottoNumberGenerateStrategy.java
@@ -1,0 +1,20 @@
+package lotto.domain.strategy;
+
+import lotto.domain.GenerateLottoNumberStrategy;
+
+import java.util.List;
+
+public class ManualLottoNumberGenerateStrategy implements GenerateLottoNumberStrategy {
+
+    private final List<Integer> lottoNumbers;
+
+    public ManualLottoNumberGenerateStrategy(List<Integer> lottoNumbers) {
+        this.lottoNumbers = lottoNumbers;
+    }
+
+    @Override
+    public List<Integer> generate() {
+        return lottoNumbers;
+    }
+
+}
