+++ src/main/java/lotto/domain/strategy/RandomLottoNumberGenerateStrategy.java
@@ -1,22 +1,27 @@
-package lotto.util;
+package lotto.domain.strategy;
+
+import lotto.domain.GenerateLottoNumberStrategy;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class RandomLottoNumberGenerator {
+public class RandomLottoNumberGenerateStrategy implements GenerateLottoNumberStrategy {
     public static final int LOTTERY_MIN_NUMBER = 1;
     public static final int LOTTERY_MAX_NUMBER = 45;
     public static final int LOTTO_NUMBER_COUNT = 6;
 
-    public static List<Integer> generate() {
+    @Override
+    public List<Integer> generate() {
         List<Integer> numbers = IntStream.rangeClosed(LOTTERY_MIN_NUMBER, LOTTERY_MAX_NUMBER)
                 .boxed()
                 .collect(Collectors.toList());
 
         Collections.shuffle(numbers);
-
-        return numbers.subList(0, LOTTO_NUMBER_COUNT);
+        return numbers.stream()
+                .limit(LOTTO_NUMBER_COUNT)
+                .sorted()
+                .collect(Collectors.toList());
     }
 }
